requirements.txt
flask
gunicorn
python-dotenv
psycopg[binary]
Pillow
pillow-heif
